+++ newton/_src/solvers/kamino/examples/sim/example_sim_dr_legs.py
@@ -264,8 +264,6 @@ class Example:
             config.solver.dvi.block_iterations = 4
             config.solver.dvi.contact_iterations = 2
             config.solver.dvi.bilateral_solve_period = 1
-            config.solver.dvi.contact_jacobi_omega = 0.45
-            config.solver.dvi.contact_jacobi_relaxation = 0.9
             config.solver.dvi.warmstart_mode = "containers"
             config.solver.dvi.contact_warmstart_method = "key_and_position_with_net_force_backup"
         config.solver.padmm.use_graph_conditionals = use_graph_conditionals